Relocate IntegerRangeCharWidget
